Check termination condition within integration step
Added support for assessing the propagation termination condition
during the evaluation of the state derivatives needed to integrate to
the next step (e.g. k1, k2, k3, k4 of the RK4 integrator). Only
included in the RK4 integrator, not in the variable step-size
integrator yet.

The condition is linked to the
propagationTerminationCondition_->checkStopCondition only for the DSST
propagator. For Cowell or Encke (or other), it is set to
boost::lambda::constant( false ), so the termination condition is only
assessed at the end of each integration step, as usual.