Custom Space Weather file
It is possible to provide a custom space weather file (with solar
activity data for the NRLMSISE00 atmosphere model) during simulation
set-up. If AtmosphereSettings is used, the default file stored in Tudat
will be used. If the new NRLMSISE00AtmosphereSettings is used, the user
can specify the path to a custom file, which should still have the same
format as in https://celestrak.com/SpaceData/sw19571001.txt.